+++ demo/app/pages/+default/home/home.component.ts
@@ -1,0 +1,28 @@
+import {Component, ChangeDetectionStrategy} from '@angular/core';
+import {ComponentRedirectRoute, ComponentRoute} from '@anglr/common/router';
+import {DebugDataCopyClickModule} from '@anglr/common/material';
+import {WithScrollableCssClass} from '@anglr/common';
+import {RenderMarkdownDirective} from '@anglr/md-help';
+
+/**
+ * Home component
+ */
+@Component(
+{
+    selector: 'home-view',
+    templateUrl: 'home.component.html',
+    standalone: true,
+    imports:
+    [
+        DebugDataCopyClickModule,
+        RenderMarkdownDirective,
+    ],
+    preserveWhitespaces: true,
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+@ComponentRedirectRoute('', 'home')
+@ComponentRoute({path: 'home'})
+@WithScrollableCssClass()
+export class HomeComponent
+{
+}