+++ highlightjs/src/index.ts
@@ -1,0 +1,1 @@
+export * from './misc/extensions/highlightjs/highlightjs.extension';