+++ src/extensions/index.ts
@@ -1,0 +1,1 @@
+export * from './assetsPathPrefix/assetsPathPrefix.extension';